Relative timing on CPU vs clock
Relative timing can now be performed relative to clock as well
as CPU.
SHVThreadBase::GetTicksInMilliSecs is relative to CPU, whereas
the new SHVTime::GetRelativeTimeInMilliSecs is relative to the
clock.
There is no functional difference between these unless the system
is suspended. The SHVThreadBase one will be suspended as well,
where the SHVTime one will include the suspended time.
Unfortunately since I couldn't find a reliable "CPU" tick count
for windows it doesn't work as intended on that platform.