SHVString overhaul
SHVString had previously been bound to the platform, meaning
on UNICODE windows it was 16bit, and otherwise it was 8 bit,
which meant iso8859-1.
This has now changed so default SHVString is UTF8, with serious
implications. The default string is no longer the same as the
native string, and characters can have variable length.